requirements.txt
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
torch==2.4.1
arch==7.1.0
scikit-learn==1.5.2
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
yfinance==0.2.46